unit tests and refactor
