+++ src/test/java/com/aol/cyclops/functions/collections/extensions/standard/ListXCollectableTest.java
@@ -4,7 +4,7 @@ import org.jooq.lambda.Collectable;
 
 import com.aol.cyclops.control.ReactiveSeq;
 import com.aol.cyclops.data.collections.extensions.standard.ListX;
-import com.aol.cyclops.react.lazy.sequenceM.CollectableTest;
+import com.aol.cyclops.react.lazy.sequence.CollectableTest;
 
 public class ListXCollectableTest extends CollectableTest {
