+++ src/test/java/com/aol/cyclops/react/lazy/sequence/LazyFutureStreamCollectableTest.java
@@ -1,9 +1,9 @@
-package com.aol.cyclops.react.lazy.sequenceM;
+package com.aol.cyclops.react.lazy.sequence;
 
 import org.jooq.lambda.Collectable;
 
 import com.aol.cyclops.control.ReactiveSeq;
-import com.aol.cyclops.react.lazy.sequenceM.CollectableTest;
+import com.aol.cyclops.react.lazy.sequence.CollectableTest;
 import com.aol.cyclops.types.futurestream.LazyFutureStream;
 
 public class LazyFutureStreamCollectableTest extends CollectableTest {
