+++ src/test/java/com/aol/cyclops/react/lazy/sequence/hotstream/SchedulingTest.java
@@ -1,4 +1,4 @@
-package com.aol.cyclops.react.lazy.sequenceM.hotstream;
+package com.aol.cyclops.react.lazy.sequence.hotstream;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
