+++ src/test/java/com/aol/cyclops/streams/ReactiveSeqCollectableTest.java
@@ -3,7 +3,7 @@ package com.aol.cyclops.streams;
 import org.jooq.lambda.Collectable;
 
 import com.aol.cyclops.control.ReactiveSeq;
-import com.aol.cyclops.react.lazy.sequenceM.CollectableTest;
+import com.aol.cyclops.react.lazy.sequence.CollectableTest;
 
 public class ReactiveSeqCollectableTest extends CollectableTest {
